moved ROS graph specific funtionality in a new module
